+++ yapyg/sprites.pxd
@@ -1,4 +1,4 @@
-# Copyright (c) 2014 Raihan Kibria
+# Copyright (c) 2015 Raihan Kibria
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -26,13 +26,13 @@ cdef int IDX_STATE_SPRITES
 
 cpdef initialize(int state_idx, list state)
 cpdef destroy(list state)
-cpdef insert(list state, str sprite_name, tuple textures, int speed, list pos_offset,
+cpdef insert(list state, str sprite_name, tuple textures, float speed, list pos_offset,
              tuple scale, int enable, list pos, int screen_relative=*, int play_once=*)
 cpdef delete(list state, str sprite_name)
 cpdef list get(list state, str sprite_name)
 cpdef tuple get_pos(list state, str sprite_name)
 cpdef set_enable(list state, str sprite_name, int enable)
 
-cdef void draw_sprite(list state, canvas, tuple view_pos, int view_scale, str sprite_name,
+cdef void draw_sprite(list state, canvas, tuple view_pos, float view_scale, str sprite_name,
                       texture, list pos, list scale, tuple origin_xy, int screen_relative)
-cdef void draw(list state, canvas, int frame_time_delta, int view_scale)
+cdef void draw(list state, canvas, float frame_time_delta, float view_scale)
